Convert test to use supported types

+++ src/test/hnetwork_test.cpp
@@ -4,8 +4,8 @@
 #include "test.h"
 #include <thread>
 
-HNetworkProcessor<int> *server;
-HNetworkProcessor<int> *client;
+HNetworkProcessor<int16_t> *server;
+HNetworkProcessor<int16_t> *client;
 
 class HNetwork_Test: public Test
 {
@@ -24,54 +24,54 @@ class HNetwork_Test: public Test
 
     private:
 
-        class TestReader : public HReader<int>
+        class TestReader : public HReader<int16_t>
         {
             private:
 
-                int* _data;
+                int16_t* _data;
                 bool _first;
 
             public:
 
-                TestReader(int* data, int blocksize):
+                TestReader(int16_t* data, int blocksize):
                     _data(data),
                     _first(true)
                 {}
 
-                int Read(int* dest, size_t blocksize)
+                int Read(int16_t* dest, size_t blocksize)
                 {
                     if (!_first)
                     {
                         return 0;
                     }
                     _first = false;
-                    memcpy(dest, _data, blocksize * sizeof(int));
+                    memcpy(dest, _data, blocksize * sizeof(int16_t));
                     return blocksize;
                 }
         };
 
-        class TestWriter : public HWriter<int>
+        class TestWriter : public HWriter<int16_t>
         {
             private:
 
-                int* _received;
+                int16_t* _received;
                 bool _first;
 
             public:
 
-                TestWriter(int* received, int blocksize):
+                TestWriter(int16_t* received, int blocksize):
                     _received(received),
                     _first(true)
                 {}
 
-                int Write(int* src, size_t blocksize)
+                int Write(int16_t* src, size_t blocksize)
                 {
                     if (!_first)
                     {
                         return 0;
                     }
                     _first = false;
-                    memcpy(_received, src, blocksize * sizeof(int));
+                    memcpy(_received, src, blocksize * sizeof(int16_t));
                     return blocksize;
                 }
         };
@@ -90,16 +90,16 @@ class HNetwork_Test: public Test
         void test_reading_server_to_writing_client()
         {
             bool terminated = false;
-            int expected[] = {1, 3, 2, 4, 3, 5, 4, 6, 5, 7, 6, 8, 7, 9};
-            int received[14];
-            memset(received, 0, 14 * sizeof(int));
+            int16_t expected[] = {1, 3, 2, 4, 3, 5, 4, 6, 5, 7, 6, 8, 7, 9};
+            int16_t received[14];
+            memset(received, 0, 14 * sizeof(int16_t));
 
             TestReader rdr(expected, 14);
-            server = new HNetworkProcessor<int>(1234, &rdr, 14, &terminated);
+            server = new HNetworkProcessor<int16_t>(1234, &rdr, 14, &terminated);
             std::thread serverThread(runServer);
             sleep(5);
             TestWriter wr(received, 14);
-            client = new HNetworkProcessor<int>("127.0.0.1", 1234, &wr, 14, &terminated);
+            client = new HNetworkProcessor<int16_t>("127.0.0.1", 1234, &wr, 14, &terminated);
             std::thread clientThread(runClient);
 
             clientThread.join();
@@ -108,22 +108,22 @@ class HNetwork_Test: public Test
             delete client;
             delete server;
 
-            ASSERT_IS_EQUAL(memcmp(received, expected, 14 * sizeof(int)), 0);
+            ASSERT_IS_EQUAL(memcmp(received, expected, 14 * sizeof(int16_t)), 0);
         }
 
         void test_reading_client_to_writing_server()
         {
             bool terminated = false;
-            int expected[] = {1, 3, 2, 4, 3, 5, 4, 6, 5, 7, 6, 8, 7, 9};
-            int received[14];
-            memset(received, 0, 14 * sizeof(int));
+            int16_t expected[] = {1, 3, 2, 4, 3, 5, 4, 6, 5, 7, 6, 8, 7, 9};
+            int16_t received[14];
+            memset(received, 0, 14 * sizeof(int16_t));
 
             TestWriter wr(received, 14);
-            server = new HNetworkProcessor<int>(1235, &wr, 14, &terminated);
+            server = new HNetworkProcessor<int16_t>(1235, &wr, 14, &terminated);
             std::thread serverThread(runServer);
             sleep(5);
             TestReader rdr(expected, 14);
-            client = new HNetworkProcessor<int>("127.0.0.1", 1235, &rdr, 14, &terminated);
+            client = new HNetworkProcessor<int16_t>("127.0.0.1", 1235, &rdr, 14, &terminated);
             std::thread clientThread(runClient);
 
             clientThread.join();
@@ -132,6 +132,6 @@ class HNetwork_Test: public Test
             delete client;
             delete server;
 
-            ASSERT_IS_EQUAL(memcmp(received, expected, 14 * sizeof(int)), 0);
+            ASSERT_IS_EQUAL(memcmp(received, expected, 14 * sizeof(int16_t)), 0);
         }
 } hnetwork_test;
